Implement WebAuthenticationHelper. Closes #2
It would be preferred to use the ResourceFilters annotation to limit the REST
requests to sys admins only, but for now we have not been able to make it work.
Instead we use this custom helper and try to make sure that it is invoked for
every REST method that has to be protected.

The WebAuthenticationException that the helper is throwing is a
RuntimeException. Sonarqube / Sonarcloud rates this as a bug but we want to
stick to the default solutions here. So, the corresponding lines are "annotated"
with a '// NOSONAR' comment.